Checkpath and compiler fixes
git-svn-id: http://svn.code.sf.net/p/scst/svn/trunk@485 d57e44dd-8a1f-0410-8b47-8ef2f437770f

+++ scst/src/dev_handlers/scst_vdisk.c
@@ -2182,7 +2182,7 @@ restart:
 		}
 		full_len -= err;
 		for (i = 0; i < e; i++) {
-			if (eiv->iov_len < err) {
+			if ((long long)eiv->iov_len < err) {
 				err -= eiv->iov_len;
 				eiv++;
 				eiv_count--;
